server/requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
google-genai
groq
pdfplumber
reportlab
plotly
bcrypt
